fix(ui): resolve ambiguous sin function type
Cast focusGlowPhase to Double and sin result to CGFloat to resolve type ambiguity

+++ PlayCoverManager/Views/QuickLauncherView.swift
@@ -895,7 +895,7 @@ private struct iOSAppIconView: View {
                                 lineWidth: 8
                             )
                             .blur(radius: 4)
-                            .opacity(0.6 + sin(focusGlowPhase) * 0.3)
+                            .opacity(0.6 + CGFloat(sin(Double(focusGlowPhase))) * 0.3)
                         
                         // Middle glow layer
                         RoundedRectangle(cornerRadius: 19)
